Change sector loader to load any sector
Previously, the sector loader in the boot sector would load based
on the FAT cluster number. It became apparent that being able to
load sectors arbitrairily would be needed. This changes the code
to be able to load any sector on the disk, and adds a function to
convert cluster numbers to sector numbers.