Switch text output to writing to video memory
The code to print to the screen used to use BIOS inturrupts to
print to the screen. Because this may not work in protected mode
when inturrupts are disabled, it may be nessesary to switch to
a method that doesn't use BIOS inturrupts. It's better in the long
run anway.